+++ src/bittytax/conv/parsers/kucoin.py
@@ -433,6 +433,7 @@ def parse_kucoin_account_history_funding(
     data_row: "DataRow", parser: DataParser, **kwargs: Unpack[ParserArgs]
 ) -> None:
     if "History_Funding" not in kwargs["filename"]:
+        # Only the Funding Account can contain airdrops
         raise DataFormatNotSupported(kwargs["filename"])
 
     row_dict = data_row.row_dict
@@ -583,7 +584,7 @@ DataParser(
     [
         "UID",
         "Account Type",
-        lambda c: re.match(r"(^Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
         "Coin",
         "Amount",
         "Fee",
@@ -602,7 +603,7 @@ DataParser(
     [
         "UID",
         "Account Type",
-        lambda c: re.match(r"(^Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
         "Remarks",
         "Status",
         "Fee",
@@ -620,7 +621,7 @@ DataParser(
     [
         "UID",
         "Account Type",
-        lambda c: re.match(r"(^Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
         "Coin",
         "Amount",
         "Fee",
@@ -659,7 +660,7 @@ DataParser(
     [
         "UID",
         "Account Type",
-        lambda c: re.match(r"(^Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
         "Remarks",
         "Status",
         "Fee",
@@ -680,7 +681,7 @@ DataParser(
         "UID",
         "Account Type",
         "Order ID",
-        lambda c: re.match(r"(^Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
         "Staked Coin",
         "Product Type",
         "Product Name",
@@ -708,11 +709,29 @@ DataParser(
         "Total Realized PNL",
         "Total Funding Fees",
         "Total Trading Fees",
-        lambda c: re.match(r"(^Position Opening Time\((UTC[-+]\d{2}:\d{2})\))", c),
-        lambda c: re.match(r"(^Position Closing Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Position Opening Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^,?Position Closing Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
     ],
     worksheet_name="Kucoin F",
     all_handler=parse_kucoin_futures,
+)
+
+DataParser(
+    ParserType.EXCHANGE,
+    "KuCoin Account History",
+    [
+        "UID",
+        "Account Type",
+        "Currency",
+        "Side",
+        "Amount",
+        "Fee",
+        lambda c: re.match(r"(^Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
+        "Remark",
+        "Type",  # New field
+    ],
+    worksheet_name="KuCoin A",
+    row_handler=parse_kucoin_account_history_funding,
 )
 
 # Account History_Funding Account (Bundle)
@@ -726,7 +745,7 @@ DataParser(
         "Side",
         "Amount",
         "Fee",
-        lambda c: re.match(r"(^Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
         "Remark",
     ],
     worksheet_name="KuCoin A",
@@ -747,7 +766,7 @@ DataParser(
         "Filled Amount",
         "Filled Volume",
         "Filled Volume (USDT)",
-        lambda c: re.match(r"(^Filled Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Filled Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
         "Fee",
         "Tax",
         "Maker/Taker",
@@ -789,7 +808,7 @@ DataParser(
         "UID",
         "Account Type",
         "Order ID",
-        lambda c: re.match(r"(^Order Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Order Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
         "Symbol",
         "Side",
         "Order Type",
@@ -799,7 +818,7 @@ DataParser(
         "Filled Amount",
         "Filled Volume",
         "Filled Volume (USDT)",
-        lambda c: re.match(r"(^Filled Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Filled Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
         "Fee",
         "Fee Currency",
         "Tax",  # New field
@@ -816,7 +835,7 @@ DataParser(
         "UID",
         "Account Type",
         "Order ID",
-        lambda c: re.match(r"(^Order Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Order Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
         "Symbol",
         "Side",
         "Order Type",
@@ -826,7 +845,7 @@ DataParser(
         "Filled Amount",
         "Filled Volume",
         "Filled Volume (USDT)",
-        lambda c: re.match(r"(^Filled Time\((UTC[-+]\d{2}:\d{2})\))", c),
+        lambda c: re.match(r"(^Filled Time\((UTC|UTC[-+]\d{2}:\d{2})\))", c),
         "Fee",
         "Fee Currency",
         "Status",