+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Kraken parser: "earn" with a fee must be a "reward" if the subtype is empty.
 - Accounting tool: in PDF reports only word-wrap for asset-table and total-table.
 - KuCoin parser: Fee can be empty for deposits and withdrawals.
+- KuCoin parser: UTC can be without an offset.
 ### Added
 - Accounting tool: new PDF report format.
 - Config: added classic_report parameter to switch to legacy PDF report format.